+++ CHANGES.rst
@@ -70,6 +70,10 @@ Changes in this release include the following:
 
 * Fix type error that would occur using pycolourchooser. (#957)
 
+* Add wrapper for wx.StaticBox.GetBordersForSizer and use it in the demo to do
+  platform-specific layout of the items in the StaticBox. (#974)
+
+
 
 
 4.0.3 "The show must go on. (Die show-stoppers! Die!)"